pyproject.toml
[project]
name = "test-fastapi-celery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.116.2",
    "uvicorn[standard]==0.35.0",
    "celery==5.4.0",
    "flower==2.0.1",
    "redis==5.0.8",
    "psycopg2-binary==2.9.9",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "typing-extensions==4.15.0",
    "langchain-openai==1.0.1",
    "langchain-experimental==0.0.42",
    "pandas==2.3.3",
    "yfinance==0.2.66",
    "praw==7.8.1",
    "feedparser==6.0.12",
    "stockstats==0.6.5",
    "eodhd==1.0.32",
    "langgraph==1.0.1",
    "chromadb==1.2.1",
    "setuptools==80.9.0",
    "backtrader==1.9.78.123",
    "akshare==1.17.71",
    "tushare==1.4.24",
    "finnhub-python==2.4.25",
    "parsel==1.10.0",
    "tqdm==4.67.1",
    "pytz==2025.2",
    "chainlit==2.8.3",
    "rich==14.2.0",
    "questionary==2.1.1",
    "langchain-anthropic==1.0.0",
    "langchain-google-genai==3.0.0",
    "streamlit==1.50.0",
    "SQLAlchemy==2.0.36",
    "alembic==1.14.0",
]
